refactor: 컴포넌트 분리 및 css 적용용
