+++ src/css/index.css
@@ -79,9 +79,27 @@ header > nav > ul > li > a {
 /* ------------ News Page-------------- */
 #NewsPage {
   position: absolute;
-  top: 64px;
+  top: 100px;
   width: 100%;
-  height: calc(100% - 68px);
+  height: auto;
+  display: flex;
+  justify-content: center;
+}
+/* ------------ Upcoming Page-------------- */
+#UpComing {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  width: 90%;
+}
+#UpComing h1 {
+  width: 500px;
+  padding: 10px 25px;
+  font-size: 1.5rem;
+  font-weight: 500;
+  color: #f1f1f1;
+  margin-bottom: 20px;
+  background-color: rgb(85, 85, 177);
 }
 /* ------------ Common -------------- */
 #Card {
@@ -89,19 +107,43 @@ header > nav > ul > li > a {
   align-items: center;
   padding: 20px;
 
-  width: 400px;
+  width: 510px;
   height: 200px;
+
+  margin-bottom: 10px;
+
+  background-color: #fafbff;
+  border-radius: 10px;
+  box-shadow: rgba(60, 64, 67, 0.3) 0px 1px 2px 0px,
+    rgba(60, 64, 67, 0.15) 0px 1px 3px 1px;
 }
 #Card .card-img {
   width: 200px;
   height: 200px;
   border-radius: 5px;
+
+  margin-right: 30px;
 }
 #Card .card-info {
   width: 100%;
+  height: 100%;
 
+  margin-top: 30px;
   display: flex;
   flex-direction: column;
-  justify-content: center;
-  align-items: center;
+  /* justify-content: flex-start;
+  align-items: flex-start; */
+}
+#Card .card-info h2 {
+  font-size: 1.2rem;
+  font-weight: bold;
+
+  margin-bottom: 40px;
+}
+#Card .card-info .card-info-date {
+  display: flex;
+}
+#Card .card-info .card-info-date div {
+  margin-right: 10px;
+  font-weight: bold;
 }
