+++ src/utils/router.js
@@ -9,9 +9,6 @@ export const routes = [
   { path: /^\/$/, component: Landing },
   { path: /^\/news$/, component: News}
 ];
-
-// 비동기 요청 사용하는 컴포넌트 렌더링 위해서 
-// async:await 사용
 export const router = async (routes, path) => {
   const component = routes
     .find(route => route.path.test(path))
@@ -20,9 +17,6 @@ export const router = async (routes, path) => {
   $('#root').innerHTML = '';
   $('#root').append(await component());
 };
-
-// [x] 현재 header a태그에 종속적인데 path 받아서도 이동 할 수 있도록 구현
-  // => 클로저 활용
 export const navigate = path => e => {
   if (path) {
     window.history.pushState({}, null, path);
